FCCustomLuaWindow mixin fixes
- Move DebugClose to FCMCustomLuaWindow to enable it to run before any handlers and make it configurable
- Move timer methods to FCXCustomLuaWindow
- Pass self as first parameter to timer handlers
